requirements.txt
# FamilyFinance - Dependencies (Slim/Optimized)
# Target: <150MB RAM startup, <512MB under load

# Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
python-multipart==0.0.9

# Cloud Storage
google-cloud-storage==2.14.0

# Image Processing (WebP compression)
pillow==10.2.0

# HTTP Client (for health checks & external APIs)
httpx==0.26.0

# Redis (sessions/cache)
redis==5.0.1

# Monitoring
prometheus-client==0.19.0

# Logging
python-json-logger==2.0.7
structlog==24.1.0

# System monitoring
psutil==5.9.8
